deactivateuser reducer
